+++ changelog.md
@@ -1,6 +1,15 @@
 # GroovyForge Changelog
 
-## [3.0.2] - 2020-02-08
+## [3.1.0] - 2020-11-23
+### Added
+### Changes
+- Merged FMLGroovyModLanguageProvider & FMLGroovyModTarget
+### Fixed
+- FMLGroovyModContainer: issues when loading Minecraft
+### Removed
+- FMLGroovyModTarget: See changes above
+
+## [3.0.2] - 2020-10-24
 ### Added
 ### Changes
 - Updated to Minecraft 1.16.3
@@ -9,7 +18,7 @@
 ### Fixed
 ### Removed
 
-## [3.0.1] - 2020-02-08
+## [3.0.1] - 2020-08-02
 ### Added
 - Initial Release for Minecraft 1.16.1
 ### Changes